build: update repository URL and Gradle version
The repository URL in the gradle.properties file was modified from 'https://github.com/nekofar/intellij-hardhat' to 'https://github.com/KartanHQ/intellij-hardhat' indicating a change in the repository's host. Also, the Gradle version was updated from 8.1.1 to 8.2.1 ensuring that we work with the latest version of Gradle. Lastly, the workaround for the Kotlin Compiler OutOfMemoryError was removed, presumably because the error has been fixed in the newer Gradle version.