chore: adjusted spacing, no functional changes in plugin.xml
The spacing in the 'plugin.xml' file has been adjusted for code readability and consistency in the project. No functional changes have been made. Changes made in the '.github/stale.yml' file seems to be minor and non-functional, possibly due to an auto-save function in the IDE. No newline added at the end of the file."