docs: update GitHub repository references in README.md
The GitHub references in the README.md file were updated to reflect the change in the project's repository location. The URLs for the workflow status badge and license badge were pointed to the new `KartanHQ/intellij-hardhat` repository instead of the old `nekofar/intellij-hardhat` repository. This ensures that the badges display the correct information from the current repository.